Fixes switching to work semi-intelligently.
When switching to a different node the frm_switch() function finds the
first match that works in the history and switches to that.

The effect is that `frame switch root/proj1` will switch to the last
used subnode in the tree `root/proj1`.

BUG: cannot switch to root anymore, because the last used node that
matches is the current one. Can be mitigated with a 'top' command.